updated blast build class
